Change how unaccounted_header_name is passed into samtools split
Disallow '-u FILE1:FILE2' usage where FILE1 was the output file
and FILE2 contained a replacement header to use when writing
FILE1.  Add a new option '-h' for the replacement header so the
new way of specifiying both files would be '-u FILE1 -h FILE2'.
The old method didn't work well on Windows due to drive letters
or with URLs.  '-h' was chosen for the new option as samtools
merge already uses it for a similar purpose.

Adds a very basic check that the new header is compatible with
the one in the input file.  This just looks for the same number
of references and checks that they are the same length.  It
doesn't bother to check the names or any other fields.  This
is mainly to prevent invalid memory accesses that could happen
if the reference name table shrinks compared to the input file.

Includes NEWS update due to changed behaviour.