Minor whitespace / formatting fixes.
Removed the ".bai" check in samtools view -X and samtools stat -X.
This enables the code to work on CRAM as well as bai files with funky
names.  Note: this may make error reporting less obvious to the users
(eg failing to load a region as an index due to missing argument).

Also added tests for -X and fixed a tiny memory leak.