Allow some header reordering in tests
Needed so that samtools tests will continue to work through the
HTSlib header API merge.

Fix addrprg headers that have spaces instead of tabs on one line.
Not sure if this was deliberate, as one of the tests seemed
to rely on the broken line being there.  As it caused a number
of the addrprg tests to start to fail, the line is fixed and the
offending test is modified to not use it.

Change the way test_count_rg and test_filter_header redirect
stderr to stop error messages from being lost.  Has to be
done by flushing and file descriptor redirection instead of
fdopen() as MinGW doesn't allow assigning to stderr.