Force header write to SAM.gz with --write-index.
Arguably a user error, but if we attempt to write to compressed SAM
without a header and also request auto-indexing, then we silently
also write the header anyway as reading back won't work without it.

Technically this is also true of non-indexed samtools so this creates
an odd difference in usage. Eg "samtools view in.bam | samtools view
-c" will give an error unless we modify the first view to be view -h.
We can't really be sure what the user wants though in that situation
and they may deliberately be avoiding the header so they can pass the
view output into a processing script, whereas we know requesting an
index implies wanting a functioning index.