Fixed invalid assertion.
b may not be set to NULL if the "b_score > score[rp]" condition later
on is not met.  I don't understand what this is attempting to do, but
given b isn't always NULL we shouldn't assert that it is.

Also updated documentation to be clearer on where name collated data matters.