Fix gcc 8.1.0 sprintf overflow warning
Ensure readable_bps() can't go past the end of the units array,
in case it ever gets fed an out of range value.

Reduce size of buf array used with readable_bps().
Increase size of buf2 array used with center_text().

(gcc complaint was that the string in buf may not fit in buf2).