Add estimate_library_size comments and change function name
Picard's estimateLibrarySize suggests it's using the
Lander-Waterman equation, but that's not quite right (L-W is about
estimating the number of gaps in an assembly, which is a slightly
different calculation.)  It's actually using the expected value
for the number of items in a set that have been covered after
a given number have been randomly selected with replacement.

For discussion, see this samtools-help thread:

https://sourceforge.net/p/samtools/mailman/samtools-help/thread/DUB405-EAS154589A1ACEF2BE4C573D4592180%40phx.gbl/

Change lw_equation to coverage_equation, and add a comment
explaining how it works.