Fix heap cleanup-on-fail in bam_merge_simple
Bam records in the heap could have come from a file or an in-memory
block.  Only the ones from files should be freed.  Fix an incorrect
assumption that the first entries all came from files.  Instead,
look in heap[i].i to see if the bam record came from a file or
an in-memory block.