Limit memory use on sort tests
So that the test harness is less likely to run out of memory on
small virtual machines.  The files being sorted are small so
there's no need for the 768Mbytes per thread that sort usually
requests.  With this change the minimum memory needed to complete
all the current tests on Linux x86_64 reduces from 2Gbytes to
about 600M.