Make hts_with_curses.m4 test if -ltinfo is needed for linking
Some symbols for ncurses (and possibly others) actually come from
libtinfo. Normally it's arranged that these are included when
linking just with -lncurses, but this can break down in some cases
(when the --no-add-needed linker option is used, and libncurses.so
is a link to the real library instead of a linker script that
adds the missing -ltinfo).

To make the tests in ax_with_curses.m4 resistant to this problem,
add a test to see if -ltinfo is needed to get the cbreak() symbol,
which is one of the ones affected.  If it is, -ltinfo gets appended
to CURSES_LIB.