Delete the "samtools import" subcommand.
This was deleted from the help / usage statements in 2009 (cc207d8)
between 0.1.5 and 0.1.6.  Thus it has never been a documented command
since the official launch of Samtools.

Import was just a wrapper around view, but the wrapper broke when we
added PG headers.  While a fix is trivial, this PR simply removes the
dead and untested code instead.

Fixes #1183 in as far as turning the segmentation fault into a handled
usage error instead.