Add samtools ampliconclip (PR #1219)
Clips the 5' end of reads based on values in a bed file.
This is useful for amplicon sequencing experiments, where
the first few bases sequenced come from the PCR primer
used, and not from the sample of interest.

Applied clipping can be soft (default) or hard.
Unclipped reads can be written out as normal, or
optionally filtered or have their QCFAIL flag set.

Writes some basic stats on number of reads clipped either
to a file, or stderr (default).

As clipping may move the start point of some reads, the
sort order in the header is set to "unknown" if it previously
was "coordinate".

Note that this is a work-in-progress.  It does not try to
adjust MD tags after clipping, and insert sizes will be
slightly incorrect.  To fix them, pass the output
through calmd and fixmate.

Co-authored-by: Rob Davies <rmd+git@sanger.ac.uk>