Change ampliconstats TLEN to TCOORD.
The original TLEN was rather unwieldy, being a non-sparse array of
2000 lengths as a histogram.  This is unwieldy and unscalable, but
also not actually that useful.  We also didn't have anything that
plotted it.

Length alone isn't as helpful as (start,end) tuples as the latter
permits us to marry them up to LEFT/RIGHT primer pairs, possibly
permitting QC on which primers work and which are just potential for
causing more problem than they're worth.

The new code uses a hash table so it can cope with any length.