Autoscale read count/depth plots.
The minimum value is now always 1, for consistency across all plots.
We also compute the maximum depth for any file and use that (rounded
up) for all plots.  This ensures easy comparison between figures in
the same run.

Note we still can't do this for the combined-reads.png file unless we
rejig the ordering of the code as the gnuplot script for that is
emitted as it parses, so we don't know the maximum until it's too
late.

(TODO: Possibly we could store the data in a gnuplot variable, still
writing as we go, and have the plot lines at the end instead of start
of the gnuplot file.)