Added VDEPTH section.
This is the "valid" depth, computed as per the the tdepth plots.
Valid here means full-length coverage across the template, and is
listed in the plots as "usable depth".  (It could also be considered
to be the minimum depth across the amplicon.)

Also fixed old calculation of FDEPTH/CDEPTH when reads extend beyond
the inner-most primer locations. This was observed in an ONT run where
the average depth was larger than the number of reads (obviously
incorrect), due to the run not having the 3' primers trimmed.

Plus the combined tdepth plot is now created in plot-ampliconstats.