First draft of a "samtools consensus" command.
This support simple frequency counting as well as a more complex
Bayesian algorithm used by Gap5 and Crumble.

We can also use the mapping quality, but this is too errorprone to
take at face value, so it is capped at low / high values so as to not
produce too many errors.

Base modifications may also be output, using a similar system to the
recently updated mpileup.

Note: there is a (undocumented on command line) het-only mode.  This
is useful for debugging by comparing to VCF files.  We're not trying
to write a caller, preferring instead to simply output what the
underlying data shows without excessive tidying and filtering, but
this is a quick way to look for ways to improve it.