Add a "samtools consensus" sub-command.
This support simple frequency counting as well as a more complex
Bayesian algorithm used by Gap5 and Crumble.

We also permit use of the mapping quality, but this is too errorprone
to take at face value, so it may be capped at low/high values or
adjusted via other means such as a localised NM figure.

The main proposal is for fasta / fastq, but there is also a simplified
pileup-ish format output.  This contains the consensus column (like
old pileup, rather than mpileup).  It's also strictly one row per
consensus position rather than one per reference coordinate (by
default).

The possible use cases for this command include:

- Generating a consensus suitable for draft assembly refinement,
  similar to ICORN, where the reads are aligned to the assembly, we
  generate a revised consensus, and then realign back to that
  consensus again.

- Generation of a consensus suitable for embedding within a file for
  purposes of storing reference-based CRAM files, without needing to
  submit a reference.  This may be suitable for more adhoc work where
  we don't yet have a final assembly we wish to submit, but still wish
  to provide aligned data.  Or alternatively where we have a very
  crude assembly produced solely for the purposes of data compression
  and the contigs are not intended to be a usable product.

- Simple consensus generation for things like pandemic tracking; eg
  as a replacement of ivar consensus.  Typically data here is so deep
  that it's easy to achieve accuracy via simple methods.

It also turns out however that it's possible to use it as a basic
rapid diploid variant caller.  (It's approx 8x faster than bcftools
for this.)  This commit does not include a way to generate VCF output,
but I do have a hacky perl script to convert from pileup to VCF.  If
this is deemed useful it could become a basic option in the tool, but
I don't think it's really the intention.  Generation of VCF was purely
as a test for evaluating the robustness of the consensus, which only
needs to be adequate for the use cases above.

If we wished to persue that evenue, it would be best as part of a
pipeline of tools, such as some simple mapq / baseq recalibration and
a local realignment step.  These are the sorts of things internally
that the top quality callers need to do to ensure accuracy.

For the curious however, here is an example for SynDip chr1, after
conversion of pileup to a noddy VCF.  SNP only as I haven't converted
indels (plus I'd expect them to be much less robust).

gatk 4:
SNP          Q>0 /  Q>=150
SNP   TP  265050 /  263351
SNP   FP    6819 /    3991
SNP   GT     891 /     785
SNP   FN    4605 /    6304

bcftools (2021?):
SNP          Q>0 /   Q>=46
SNP   TP  265363 /  263345
SNP   FP    5666 /    3453
SNP   GT     676 /     505
SNP   FN    4292 /    6310

freebayes (2018):
SNP          Q>0 /   Q>=12
SNP   TP  264313 /  263355
SNP   FP    6018 /    5468
SNP   GT    1673 /    1637
SNP   FN    5342 /    6300

Samtools consensus
SNP          Q>0 /   Q>=51
SNP   TP  265792 /  263377
SNP   FP    5393 /    3155
SNP   GT     686 /     342
SNP   FN    3863 /    6278

Q>=x is with bcftools filter of QUAL>=x.  It differs per tool as I
tried to normalise for the same FN figure of approx 6300, so we can
then judge how TP, FP and GT errors change.  GT is genotype error,
where the call is correct, but mislabelled as het / hom.

The above was 45x.  At 15x it's quite a bit poorer than other tools.
Similarly GIAB is a more mixed bag.  Ie this works sometimes, but it's
a robust caller.  So use with caution!