Add -e FILTER to samtools view.
This is a validation of the sam_passes_filter function, meaning it can
be used in conjunction with -U to output the reads that do not match
the filter.

Infact both can be used together.  For example

    samtools view \
        --input-fmt-option "filter=rlen-qlen >= 10" \
        -e 'ncigar <= 10' in.bam -o out.sam -U outU.sam

will pre-filter to only reads where the length on the reference is at
least 10 bases longer than the length of the sequence.  Those that
pass arr then written to either out.sam or outU.sam depending on
whether they have <= 10 cigar operations.