Disable travis tests
travis.org is closing down, and cirrus-ci works as a replacement.

Renames the .travis directory to vendor-neutral .ci_helpers.
Changes the travis badge in README.md to a cirrus-ci one.

Unfortunately this removes the only big-endian platform (s390x).
It will be substituted in the future if an alternative can be
found, but meanwhile big-endian tests will need to be run manually.