Expand "samtools --version" and add "samtools version" synonym
Expand "samtools --version" with more details and added a synonym of
"samtools version" too so it's simpler to list in the usage.

Also added the missing help subcommand in the usage list.  Note this
isn't exactly the same as samtools without args, as that is an error,
while this reports to stdout.