Add a check for multi-ref BED files in ampliconclip & ampliconstats. (PR #1398)
Long term we wish to make these cope, but for now we take a more
defensive approach by recognising the case where we would give
incorrect results.

This extends bed_pair_list_t with a fixed length buffer, which is a
temporary fudge.  It'll need something like the pooled string
allocator and a ref name pointer within bed_pair_t to properly deal
with multi-ref BEDs.

Ampliconstats now also has code to filter out any reads that are for
another chromosome, so if we have a multi-ref BAM and a single-ref BED
then it'll still work correctly.