Fix negative depth values in ampliconstats output.
These come about due to primer-dimers where the sequence start
coordinate matches a known primer, but the sequence doesn't extend
beyond the primer itself.  If the primer in question is one of a pair
(ie having an ALT primer), then the primer-dimer sequences can finish
multiple bases before the inner start coordinate for the amplicon.

This case wasn't handled, leading to a negative increment to the
number of amplicon bases and consequential negative CDEPTH.

Also applied a minor rounding error correction to CDEPTH STDDEV where
a value of zero is sometimes a tiny negative value, giving rise to
-nan when passed into sqrt.