Stop samtools cat from outputting multiple CRAM EOF markers.
Samtools cat naively copies all incoming containers, and then writes
an empty EOF container when closing the file.  This leads to multiple
internal EOF markers as they're implemented in CRAM as an empty
container.

Oddly there was code explicitly dealing with this, which seems
redundant.  Perhaps it served a purpose in early versions.  Regardless
internal EOF markers is permitted and are ignored, so this change is
just for tidyness sake and to improve robustness incase bad
implementations stop on the first EOF.