Make coverage ASCII mode true ASCII.
The example in the man page used "#" characters, but in source code it
was using a more limited UTF8 encoding of only full and half-height
blocks instead of 8 height levels.

It's a misnomer to claim this is ASCII.  We could still have a
half-height UTF8 mode, but I'm not sure I see the point in catering to
the middle case much.  ASCII however should be purely printable
without having to worry about text editors or pagers escaping
non-printable characters.  This switches to : and . chars.