Totally rewrite the "samtools depth" command.
This uses a circular buffer to accumulate and report depth, instead of
relying on mpileup.  This means it's considerably faster (particularly
on deep data and on CRAM files) and doesn't have the need for a depth
limit to avoid high memory usage.  It still silently accepts the -d/-m
(max depth) parameter, but ignores it.

All output should be identical to the old command except potentially
for the overlap removal (-s) option.  The old code here is complex,
requiring all in-flight records to be held in memory by design.
This is explicitly what I wanted to remove as it is remembering these
records-in-flight that gives the high memory usage and the requirement
of a maximum depth.

Therefore the new depth algorithm may give different results:

- It doesn't force a -q1 or higher.  This is needed with mpileup as
  mpileup doesn't actually removal overlaps, it merely marks them.
  (Try old "depth -q0 -s" and "depth -s -q0" to see this.)

- Even if the user manually adds the -q1 option, the overlap removal
  is purely coordinate based.  The mpileup "removal" is coordinate,
  sequence and quality based.  If overlapping bases agree, the quality
  used is the summation of both qualities, else it's 80% of the
  highest quality.

  This gives differences where read 1 has zero qual and read 2
  overlapping has non-zero qual.  With -q1 we'd ignore that first base
  while mpileup would still count that depth.

- Overlap removal applies to all paired reads, not only those with the
  proper-pair flag set.  This is related to #1036.  Strongly
  overlapping reads are sometimes not labelled as a proper pair as
  they fall outside the expected insert size distribution, but those
  are the reads we really ought to be deduplicating.  (Note this is
  still an issue for mpileup though so it doesn't fix that issue.)

In the process of assessing this, two bugs were discovered in the old
depth code.

- Specifying multiple BAM files where the first file consists solely
  of unmapped reads produces no output.  It should report the data for
  the 2nd file onwards still.

- Due to a mistake in the CRAM_OPT_REQUIRED_FIELDS initialisation, the
  overlap removal (-s) option didn't work correctly on CRAMs, wrongly
  removing every other read regardless of read name.

Regarding samtools depth documentation, the only things that really
changes here are -d/-m options are now ignored, and the overlap
removal has a different algorithm.

I haven't put it in the man page, but for reference

    samtools depth -J -s in.bam

matches

    samtools mpileup -Q0 in.bam | cut -f 1,2,4

There is no anti-J (skip deletions) option for mpileup, so we cannot
perfectly simulate the old samtools depth without parsing the SEQ
column, which sadly cannot be done using traditional regexp hackery
either so it's a complex operation.