Fix depth expected output.
The difference here is caused through use of quality -1.
The old depth forces -q1 when we enable overlap removal.  If we do
that here then the changes based 51-55 are correct, but we get new
differences at 77-80.

The key differences between old overlap removal and new overlap
removal is the new one is simply adjusting the second read to start
where the first read finishes, to avoid overlap, while the old method
modifies quality values to be the maximum of the two reads (if seq
agrees) and 80% of the highest quality value if they disagree; with
the second duplicate read getting qual 0.

This complex behaviour is hard to achieve short of using mpileup again.

The other difference is that depth overlap removal now considers all
paired reads and not only thise with the properly-paired flag set.