Add alternative `samtools merge -o FILE` option
The transition of `samtools sort [-o] in.bam out.prefix` to the typical
Unix-style `samtools sort [-o out.bam] in.bam` is long since complete.

It was intended to transition `samtools merge out.bam in1.bam in2.bam...`
similarly, but as this was comparatively easy (no conflicting meaning
for -o to deal with) it never quite happened.

Add support for `samtools merge -o out.bam [options] in1.bam in2.bam...`,
which is easy to support alongside the existing syntax. The difficult
transition here would be to default to writing to standard output, and
this PR does not attempt to do that. As the existing syntax is quite
entrenched, it would not be feasible to move to that more typical
default without a long transition period, so for now `-o -` remains
required in pipelines.