Add consensus show-del and show-ins options.
The default is deletions are omitted (no "*") and insertions shown,
making it a true consensus.  If you want a reference-coord consensus
(eg for an embedded CRAM reference) then use
    samtools consensus -show-del=y --show-ins=n -a FILE

TODO:
- Use quality values properly (it's just counting always currently).
- A better strategy in the bayesian "gap5" mode.  For now it just
  computes the insertions in the same manner, but it ought to honour
  the other parameters used by this algorithm such as mapping quality.

Ideally we'd just hack the pileup algorithm itself to return
*consensus* column by column rather than *reference* column by column,
then no change is needed to either algorithm, but that may be a
herculean task.