Flush standard output samFiles before printing error messages
When output is to the terminal (common when e.g. `samtools view` is
used to view SAM), flush it to ensure a fatal error message is not
obscured by buffered output that is printed later when the output
stream is closed. (Or in general, to ensure that diagnostics appear
interleaved in the correct place in the output.)

The print_error()/etc routines already flush stdout; add utility
functions to register a standard output samFile (really an htsFile)
that should also be flushed before printing a diagnostic message.

Register the output samFile (or vcfFile in the case of `mpileup`) in
a few commands that may be often used with stdout, notably `view`.
Convert some "read failure" error messages to use print_error() so
that they will benefit from this flushing.