Make sort better handle temporary file collisions
Make worker() try alternate temporary file names if its first
choice already exists.  As the name is no longer predictable,
it gets passed back and stored so the merge code knows which files
to open.

Improves handling of failures so it's more likely that temporary
files will be cleaned up.