Strip index name from prefix used for sort temporary files
By default sort uses the output file name as a prefix for
its temporary file names.  HTSlib also allows the name of
an index file to be appended to the output file name after
an '##idx##' delimiter.  When this has been done the delimiter
and index name part need to be removed to get the correct
prefix, especially when the index part includes path elements.