Use sam.gz for temporaries when sorting files with long references
Reference positions higher that INT32_MAX can't be stored in BAM
or CRAM < 4.0, so switch to using .sam.gz in temporary files
when the header indicates that a long reference is in use.  We
could use CRAM4 instead here, but HTSlib will currently emit
warnings about it not being standardised yet so it's less alarming
for the end user if we use sam.gz.

As a bonus, if a long reference is found, check that the output
format will be compatible.  This means sort can exit with an error
immediately, instead of waiting for the output to be opened which
may happen quite a long time after the program starts.