Fix missing test_cmd() out parameter
Adds an `out` parameter with an empty to a bam2fq test_cmd()
call that didn't have one (it did have out_map).  This is
necessary because test_cmd() always attempts to build a path
with this parameter and open it, so without a value it attempted
to open a directory.  Somehow this worked on Linux, Windows and
MacOS, but it failed on NetBSD.

Also change the validation in test_cmd() so `out` must be
present, even if `out_map` is in use.