Cull the long deprecated VCF and BCF output from mpileup.
This was marked as deprecated in 1.9, with both the usage statement
and documentation components having been removed already.

We still need bam2bcf.[ch] for the slimmest of reasons: to report
ambiguity codes in tview consensus column.  However the bulk of the
code has been trimmed to their minimum.  Frankly a basic summing
algorithm is probably good enough for this modest use case, but for
now it's left in-situ.