Code tidy up.
Improved license comments.

Added an export to pileup_seq in a header file so both definition and
call can check against the same prototype.  This was added to
bam_plbuf.h as it's related code and avoids creation of a new
bam_plcmd.h file for a single function declaration.

Cull some of the unused bit of the gap5 consensus algorithm.  In
theory we could check read groups for platform fields and select
instrument specific undercall / overcall priors, but for now we have a
fixed value.

Also minor improvements to readability and ordering.