Fixes the documentation for mpileup --output-BP-5.
The original intention of this was to modify the orientation of the
--output-BP field, as originally requested.  Due to lack of
research into the original code, sadly this was moved to the end of
the list of columns.

It was decided that we cannot fix this as it would become incompatible
with the previous release, so it's now been documented.

However given it's now its own column instead of modifying the
existing column, --output-BP-5 is no longer mutually exclusive with
--output-BP.

Fixes #1534 (albeit not in the ideal way)