Split aux tag adjustment from read filtering
Pull the part that implements the `--keep-tags` and `--remove-tags`
options out of process_aln() and put it in a new adjust_tags()
function.  This means the tag adjustment only has to be done when
actually needed - so for example it can be skipped in counting
mode.

Invalid aux tags detected in adjust_tags() are now reported
immediately as errors.  This just means they're trapped a bit
earlier, as before they would have caused the writer to fail.

Note that as this code used to come after all the process_aln()
filters had run, the tags were only ever edited if the read
passed filtering.  For compatibility the new version has been
made to work the same way, and notes have been added to the manual
page stating that the options only affect passed reads.