Optical Duplicate Marking: Add regex options for custom coordinates. (PR #1558)
For the case of non standard read names (QNAME), add options to read the coordinates and, optionally, another part of the string to test for optical duplication.

Options are:

--read-coords REGEX
This takes a POSIX regular expression for at least x and y. but can include another part of the read name to test for equality, eg lane:tile elements.

--coords-order ORDER
The order of the elements captured in the regular expression.  Default is txy where t is a part of the read name selected for string comparison and x/y the coordinates used for optical duplicate detection.

An example to match the way read name reading currently works would be:
--read-coords '(\w+:[0-9]+:[0-9]+:[0-9]+:[0-9]+):([0-9]+):([0-9]+)'
--coords-order txy
To match the read format machine:run:flowcell:lane:tile:x:y

An alternative might be:
--read-coords '^([[:digit:]]):([[:digit:]])'
--coords-order xy
To match the read format x:y:randomstuff

Also removed an unused option in the help and a bug in the duplication chain code when the coordinates fail.