Improve automatic file type detection with view -u / -1.
Previously -u and -1 were documented to also enable BAM mode.
When they were written this made sense.  There were two choices -
uncompressed SAM or BAM (uncompressed or otherwise).  Specifying the
compression level inevitably meant BAM.

Since then we've gained the ability to output SAM.gz and CRAM.  Plus
we've also gained automatic file type selection by output filename.
Hence it really doesn't make much sense for

    samtools view -o out.cram -1 in.bam

to output a BAM file named out.cram.

To ensure compatibility, the default output format is still SAM, or
BAM when -1 / -u are used, but if and only if a filename is used and
it doesn't indicate otherwise.

In the process of doing this I also fixed some bug in --write-index
when used in conjunction with -U.