Added --min-BQ and --min-MQ options to depth.
These match the equivalent long options found in samtools mpileup,
which gives a consistent way of specifying the base and mapping
quality filters.  This is in contrast to the short options, -q and -Q,
which are reversed between the two subcommands (historic).

Also updated the rarely used phase and bedcov commands, with --min-BQ
and/or --min-MQ as appropriate.

Fixes #1580