Consensus * vs N tweak.
The Bayesian consensus mode can call both homozygous (cons.call) and
heterozygous (cons.het_call) answers, although the quality score is
assigned from the heterozygous case.  So if it thinks it's a het
(cons.het_logodd > 0) then the phred score for the homozygous base is
probably zero (P_correct < 0.1).

If we're not enabling ambiguity codes in the consensus, then the "cq <
opts->cons_cutoff" check is almost always true and the base gets
converted to "N".  This doesn't make sense however as we don't have a
good traditional interpretation of "N" meaning base vs not-base.
The consensus will still get called a as pure base here (or gap),
leaving the N character only for the more traditional ACGT vs ACGT
conflict.