Allow multiple alignment files to be indexed at once
Accept either `FILE [FILE]...` or `FILE [OUTINDEX]`; when exactly two
filenames are given, interpret the second as OUTINDEX if it does not
exist or if it already contains index data. In particular, do not
interpret an existing BAM/CRAM file as an output index filename!

Add optional `-o OUTINDEX` option that can be used instead of OUTINDEX,
to align this subcommand with others that use -o to specify output files.
At present this can only be used when 1 input file is given. In future, it
could include expansion metacharacters similarly to the split subcommand.

In principle, we could use a thread pool to index multiple input files
at once (as well as multi-threading the input reading). This would be
a larger change and, at least initially, probably not worthwhile.