Add a "samtools reference" command.
This outputs a reference from a BAM or CRAM file using the MD:Z aux
tags, or direct from a CRAM file if it has been created with an
embedded reference.  The reference will inevitably have holes as it is
dependent on being covered by sequence data.

Note this still isn't going to recover a CRAM file with an absent
reference, unless the CRAM file was produced using verbatim MD:Z
instead of auto-generated ones (which is possible, but very
unlikely).