Fix template-coordinate library comparison
The bam_get_library() function returns a pointer to a static
buffer, which means the data pointed to is overwritten each
time it gets called.  This means it can't be used by
template_coordinate_key(), which wants to store pointers to
different library names.

This adds code to build a RG to library lookup table from
the header when sorting in template-coordinate mode.  The stored
strings are then safe for template_coordinate_key() to use when
storing library names.