Rename mpileup --ignore-overlaps option.
This is now --ignore-overlaps-removal or --disable-overlap-removal.
The previous name is ambiguous and is often read as an option to
enable removal of overlapping bases, while in reality this is on by
default and the option turns off the abiltiy to remove overlapping
bases.

Note the addition of "-removal" and retaining of overlaps plural is to
permit the existing "--ignore-overlaps" option to continue working as
GNU getopt long has always permitted shortened versions of long options.

Also added --disable-overlap-removal as an (IMO) better synonym, but I
am open to removing this if it is considered unnecessary.

Fixes #1663