Fix missing @RG headers when using 'samtools merge -r'
Caused by a broken test for an empty klist.  Fix by checking
the size stored in the klist struct.  It's not clear if it's
an officially supported way of testing for an empty klist
(there is currently no macro to expose size), but it works for
now.

Adds a test, as the case of no @RG headers in the inputs was
previously not covered.