Add -M option required to activate multiple input file indexing
If a user of an old samtools version used `samtools index a.bam b.bam`
after reading of this facility, they would lose b.bam's data.

To avoid the possibility of encouraging users of old versions to make
this mistake, require users of new versions to add `-M` to activate the
new facility. (Using `samtools index -M a.bam b.bam` with old samtools
would lead to an unknown option error rather than overwriting b.bam.)

In future when affected old versions have receded into the past,
we may be able to make `-M` a no-op, similar to `view -S` which is
ignored as it is no longer necessary.