Add a bam_sanitize function to fixmate and view.
The original fixmate already did some basic sanitization, such as
checking for ref "*" or pos <=0 to indicate unmapped data, but it only
did so sporadically (e.g. not for secondary alignments).  It also
treated reads overhanging the end of the reference by a single base as
worthy as marking as unmapped, but it didn't clear things like CIGAR
or MAPQ.

This is more consistent and sane interface, which can turn reads
slightly overhanging the chromosome into soft-clipped alignments.

Samtools view uses the same code, but has the sanitizing disabled by
default.  (Note the -s option is taken on view but -z is free, so I
went with the Oxford spelling of sanitize.)

The rules are:

    RNAME *           => unmapped
    POS 0             => unmapped
    POS > ref len     => unmapped
    End pos > ref_len => soft-clip
    Unmapped          => MQUAL 0, CIGAR *
    Unmapped          => drop NM, MD, CG and SM tags.