Make sort merge in memory before writing temporary files
[Rest of commit msg by jkb] The net effect of this is two fold.

1. Some of the CPU from the merging phase is moved into the sorting
   phase.  This is CPU that needed to be done anyway, but it's now
   executed before writing out the temporary files instead of when
   reading them back.

2. The number of files is dramatically reduced, being divided by the
   number of running threads.

This latter part is key for boosting performance when the total file
size being processed is many times bigger than the memory block size.
It was not uncommon to be merging 100s if not 1000s of temporary
files.  This can have a significant negative impact on filesystem
churn and disk thrashing.  While not a full solution, merging before
writing moves the point of failure further away.

Tests used 16 threads with 2GB each and Lustre for input,
output and tempory files. Sorting the 120GB Illumina NA12878 Platinum
Genomes NA12878 file from name to position order, improved from 14+
hours to under 3 hours.   Sorting it back from name to position
improved from ~3 hours to 1h16m. (NB benchmarks in conjunction with
the following buffer size commit)

On small files, the difference is not significant either way.