Improve samtools sort on lustre.
The Lustre filesystem doesn't like many short reads.  The filesystem
stat indicates the ideal buffer size is 4MB.  This code uses 1MB as a
middle ground between using excessive memory while still reducing many
small reads and writes.  The effect on other file systems tested seems
minimal.

Also modified the is_digit function used in strnum_cmp.  The
difference (tested on Intel) was around 10% less CPU in strnum_cmp.
This may not sound significant, but the heapsort function that uses
this is running in the main thread and rapidly becomes the bottleneck
once 4+ threads are used.

I assume this relies heavily on branch prediction algorithms, as
naively we'd expect a table lookup to be faster, but on this quite old
Xeon E5-2660 the conditional was quicker.  This hasn't been
benchmarked on ARM.