Make consensus -m simple use a more usual definition of min_depth.
It was previously strictly the minimum depth of data agreeing with the
hypothesised consensus call, so AAAAC is depth 4 (A) rather than 5.
Similarly AACCG in het mode was also depth 4 (AA/CC).  Now depth is
the minimum depth of all bases passing the other filtering criteria,
such as minimum mapping quality, irrespective of their contribution to
the consensus call.

Fixes #1700