Also add a --min-BQ option to samtools consensus.
The code to filter against base quality was already there, but for
some reason was never given a command-line option to modify it.

I'm unsure if it's worth while (and perhaps it got removed because it
didn't help), but it's in samtools depth so for completeness we may as
well expose it with the default being no filtering.