Fix stats breakage on long deletions when given a reference
Ensure that the length of any deletions is taken into account
when refilling the reference buffer, so it can be made long
enough.  Failing to do this caused samtools stats to report an
error on deletions longer than about 20kb, because the
count_mismatches_per_cycle() function ran off the end of the
available reference.

Fixes #1707 (Very long deletions break samtools stats)