Add depth --excl-flags synonym for -G.
This matches the behaviour of "view --excl-flags", and is similar to
"mpileup --excl-flags".  The difference is that mpileup SETs the flags
while depth/view OR additional fields onto the flags.  (Depth also has
-g to AND-off the default exclude flags, which is not needed in view
as the default is zero.) Also clarify -g in the usage indicating this
is still the code to filter-out rather than filter-in and avoid
confusion over "view --incl-flags".

This commit does not add a --incl-flags option.

Also removed a duplicate -H usage option.

Fixes #1702